+++ sketch-compose-core/src/commonMain/kotlin/com/github/panpf/sketch/state/ErrorStateImage.compose.kt
@@ -2,25 +2,25 @@ package com.github.panpf.sketch.state
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
-import com.github.panpf.sketch.state.ErrorStateImage.Builder
-import com.github.panpf.sketch.state.ErrorStateImage.Condition
 
 /**
  * Create an ErrorStateImage
+ *
+ * [configBlock] must be inline so that the status used internally will be correctly monitored and updated.
  */
 @Composable
-fun ComposableErrorStateImage(
+inline fun ComposableErrorStateImage(
     defaultImage: StateImage? = null,
-    configBlock: @Composable (Builder.() -> Unit)? = null
-): ErrorStateImage = Builder(defaultImage).apply {
-    configBlock?.invoke(this)
+    crossinline configBlock: @Composable (ErrorStateImage.Builder.() -> Unit)
+): ErrorStateImage = ErrorStateImage.Builder(defaultImage).apply {
+    configBlock.invoke(this)
 }.build()
 
 /**
  * Add a custom error state
  */
 fun ErrorStateImage.Builder.addState(
-    condition: Condition,
+    condition: ErrorStateImage.Condition,
     color: Color
 ): ErrorStateImage.Builder = apply {
     addState(condition, ColorPainterStateImage(color))